pyproject.toml
[project]
name = "T20-World-Cup-Winner-Prediction-using-ML"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.45.2",
    "numpy>=2.2.2",
    "openai>=1.61.1",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "plotly-express>=0.4.1",
    "plotly>=6.0.0",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.3",
    "scikit-learn>=1.6.1",
    "streamlit>=1.42.0",
    "trafilatura>=2.0.0",
    "twilio>=9.4.4",
    "xgboost>=2.1.4",
    "python-dotenv>=1.0.1",
]
